Refactor waveform factory function
