+++ libraries/designsystem/src/main/kotlin/io/element/android/libraries/designsystem/components/media/FakeWaveformFactory.kt
@@ -20,17 +20,15 @@ import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toPersistentList
 import kotlin.random.Random
 
-object FakeWaveformFactory {
-    /**
-     * Generate a waveform for testing purposes.
-     *
-     * The waveform is a list of floats between 0 and 1.
-     *
-     * @param length The length of the waveform.
-     */
-    fun createFakeWaveform(length: Int = 1000): ImmutableList<Float> {
-        val random = Random(seed = 2)
-        return List(length) { random.nextFloat() }
-            .toPersistentList()
-    }
+/**
+ * Generate a waveform for testing purposes.
+ *
+ * The waveform is a list of floats between 0 and 1.
+ *
+ * @param length The length of the waveform.
+ */
+fun createFakeWaveform(length: Int = 1000): ImmutableList<Float> {
+    val random = Random(seed = 2)
+    return List(length) { random.nextFloat() }
+        .toPersistentList()
 }
